+++ modules/Accounting/functions.inc.php
@@ -39,7 +39,7 @@ function _makeIncomesTextInput( $value, $column )
 
 	if ( $column === 'AMOUNT' )
 	{
-		$extra = 'size=5 maxlength=10';
+		$extra = ' type="number" step="any"';
 	}
 	elseif ( ! $value )
 	{
@@ -106,7 +106,7 @@ function _makePaymentsTextInput( $value, $name )
 
 	if ( $name === 'AMOUNT' )
 	{
-		$extra = 'size=5 maxlength=10';
+		$extra = ' type="number" step="any"';
 	}
 	elseif ( ! $value )
 	{